May 31, 2026, 2:00 PM

+++ set.h
@@ -89,5 +89,5 @@ public:
 private:
   T* m_values;
   int m_size;
-  int m_capacity;
+  size_t m_capacity;
 };
